+++ src/store/auth.ts
@@ -1,11 +1,12 @@
-import { execGitCredentialFill } from "@abstractions/gitCredentialFill";
-import { performSignInFlow } from "@abstractions/signIn";
-import * as keytarType from "keytar";
-import { commands, window } from "vscode";
+import {
+  authentication,
+  AuthenticationSession,
+  commands,
+  window
+} from "vscode";
 import { store } from ".";
 import * as config from "../config";
 import { EXTENSION_NAME } from "../constants";
-import { log } from "../logger";
 import { refreshGists } from "./actions";
 const GitHub = require("github-base");
 
@@ -13,27 +14,10 @@ export function getCurrentUser() {
   return store.login;
 }
 
-export type Keytar = {
-  getPassword: typeof keytarType["getPassword"];
-  setPassword: typeof keytarType["setPassword"];
-  deletePassword: typeof keytarType["deletePassword"];
-};
-
-function getNativeKeytar(): Keytar {
-  const vscodeRequire = eval("require");
-  return vscodeRequire("keytar");
-}
-
-const keytar = getNativeKeytar();
-
-const ACCOUNT = "gist-token";
-const SERVICE = `vscode-${EXTENSION_NAME}`;
-
 const STATE_CONTEXT_KEY = `${EXTENSION_NAME}:state`;
 const STATE_SIGNED_IN = "SignedIn";
 const STATE_SIGNED_OUT = "SignedOut";
 
-const SCOPE_HEADER = "x-oauth-scopes";
 const GIST_SCOPE = "gist";
 const REPO_SCOPE = "repo";
 const DELETE_REPO_SCOPE = "delete_repo";
@@ -45,64 +29,10 @@ export async function getApi(newToken?: string) {
   return new GitHub({ apiurl, token });
 }
 
-async function testToken(token: string) {
-  const github = await getApi(token);
-
-  try {
-    const response = await github.get("/user");
-
-    const scopeHeaderIndex = response.rawHeaders.findIndex(
-      (item: string) => item.toLowerCase() === SCOPE_HEADER
-    );
-    if (scopeHeaderIndex === -1) {
-      log.info(`Token test failed: No scopes header`);
-      return false;
-    }
-
-    const tokenScopes = response.rawHeaders[scopeHeaderIndex + 1];
-    if (!tokenScopes.includes(GIST_SCOPE)) {
-      log.info(`Token test failed: Scopes don't include gist`);
-      return false;
-    }
-
-    store.login = response.body.login;
-    store.canCreateRepos = tokenScopes.includes(REPO_SCOPE);
-    store.canDeleteRepos = tokenScopes.includes(DELETE_REPO_SCOPE);
-
-    return true;
-  } catch (e) {
-    log.info(`Token test failed: ${e.message}`);
-    return false;
-  }
-}
-
-// TODO: Support SSO when using username and password vs. just a token
-async function attemptGitLogin(): Promise<boolean> {
-  const gitSSO = config.get("gitSSO");
-  if (!gitSSO) {
-    log.info("Git SSO disabled");
-    return false;
-  }
-
-  try {
-    const token = execGitCredentialFill();
-    if (token && token.length > 0 && (await testToken(token))) {
-      log.info("Git SSO succeeded");
-      await keytar.setPassword(SERVICE, ACCOUNT, token);
-      return true;
-    }
-
-    return false;
-  } catch (e) {
-    log.info(`Git SSO failed: ${e.nessage}`);
-    return false;
-  }
-}
-
 const TOKEN_RESPONSE = "Enter token";
 export async function ensureAuthenticated() {
-  const password = await getToken();
-  if (password) {
+  const token = await getToken();
+  if (token) {
     return;
   }
 
@@ -115,38 +45,35 @@ export async function ensureAuthenticated() {
   }
 }
 
-async function deleteToken() {
-  await keytar.deletePassword(SERVICE, ACCOUNT);
+async function getSession(promptUser: boolean = false) {
+  return await authentication.getSession(
+    "github",
+    [GIST_SCOPE, REPO_SCOPE, DELETE_REPO_SCOPE],
+    { createIfNone: promptUser }
+  );
 }
 
 export async function getToken() {
-  const token = await keytar.getPassword(SERVICE, ACCOUNT);
-  return token;
+  return store.token;
+}
+
+async function attemptSignin() {
+  const session = await getSession();
+  if (session) {
+    await markUserAsSignedIn(session);
+  }
 }
 
 export async function initializeAuth() {
   markUserAsSignedOut();
 
-  const isSignedIn = await isAuthenticated();
-  if (isSignedIn) {
-    log.info("Signed in, checking for token's validity...");
-    const currentToken = (await getToken())!;
-    const tokenStillValid = await testToken(currentToken);
-    if (!tokenStillValid) {
-      log.info("Clearing token, since it is no longer valud");
-      await deleteToken();
-      return;
+  authentication.onDidChangeSessions(async (e) => {
+    if (!store.isSignedIn && e.provider.id === "github") {
+      await attemptSignin();
     }
-  } else {
-    log.info("Not signed in, attempting git SSO...");
-    const gitSSO = await attemptGitLogin();
-    if (!gitSSO) {
-      return;
-    }
-  }
+  });
 
-  log.info("Marking user as signed in");
-  await markUserAsSignedIn();
+  await attemptSignin();
 }
 
 export async function isAuthenticated() {
@@ -154,29 +81,26 @@ export async function isAuthenticated() {
   return token !== null;
 }
 
-async function markUserAsSignedIn() {
+async function markUserAsSignedIn(session: AuthenticationSession) {
   store.isSignedIn = true;
+  store.token = session.accessToken;
+  store.login = session.account.label;
+  store.canCreateRepos = session.scopes.includes(REPO_SCOPE);
+  store.canDeleteRepos = session.scopes.includes(DELETE_REPO_SCOPE);
+
   commands.executeCommand("setContext", STATE_CONTEXT_KEY, STATE_SIGNED_IN);
 
   await refreshGists();
 }
 
 export async function signIn() {
-  const token = await performSignInFlow();
+  const session = await getSession(true);
 
-  if (token) {
-    if (await testToken(token)) {
-      window.showInformationMessage(
-        "You're successfully signed in and can now manage your GitHub gists and repositories!"
-      );
-
-      await keytar.setPassword(SERVICE, ACCOUNT, token);
-      await markUserAsSignedIn();
-    } else {
-      window.showErrorMessage(
-        "The specified token isn't valid or doesn't inlcude the gist scope. Please check it and try again."
-      );
-    }
+  if (session) {
+    window.showInformationMessage(
+      "You're successfully signed in and can now manage your GitHub gists and repositories!"
+    );
+    await markUserAsSignedIn(session);
   }
 }
 
@@ -184,9 +108,4 @@ function markUserAsSignedOut() {
   store.login = "";
   store.isSignedIn = false;
   commands.executeCommand("setContext", STATE_CONTEXT_KEY, STATE_SIGNED_OUT);
-}
-
-export async function signout() {
-  await deleteToken();
-  markUserAsSignedOut();
 }